added HA icon + icons in app menu

+++ mysetup.py
@@ -303,32 +303,68 @@ def download_and_run(url, local_file_name):
 def make_icon():
     import fileinput
     import shutil
-
+    
+    print "path to script:"
     print path_to_script
-
+    lisa_path = os.path.abspath(path_to_script)
 
     in_path = os.path.join(path_to_script, "applications/lisa.desktop.in")
-    print in_path
+    in_path_ha = os.path.join(path_to_script, "applications/ha.desktop.in")
+    print "icon input path:"
+    print in_path, in_path_ha
+    
     home_path = os.path.expanduser('~')
+    
     if os.path.exists(os.path.join(home_path, 'Desktop')):
         desktop_path = os.path.join(home_path, 'Desktop')
     elif os.path.exists(os.path.join(home_path, 'Plocha')):
         desktop_path = os.path.join(home_path, 'Plocha')
     else:
         print "Cannot find desktop directory"
-    out_path = os.path.join(desktop_path, "lisa.desktop")
-    shutil.copy2(in_path, out_path)
+        desktop_path = None
+    
+    # copy desktop files to desktop
+    if desktop_path is not None:
+        out_path = os.path.join(desktop_path, "lisa.desktop")
+        shutil.copy2(in_path, out_path)
+        
+        out_path_ha = os.path.join(desktop_path, "ha.desktop")
+        shutil.copy2(in_path_ha, out_path_ha)
 
-    lisa_path = os.path.abspath(path_to_script)
+        #fi = fileinput.input(out_path, inplace=True)
+        print "icon output path:"
+        print out_path, out_path_ha
 
-    #fi = fileinput.input(out_path, inplace=True)
-    print out_path
+        for line in fileinput.input(out_path, inplace=True):
+            #coma on end makes no linebreak
+            print line.replace("@{LISA_PATH}", lisa_path),
+            
+        for line in fileinput.input(out_path_ha, inplace=True):
+            #coma on end makes no linebreak
+            print line.replace("@{LISA_PATH}", lisa_path),
+            
+    # copy desktop files to $HOME/.local/share/applications/
+    # to be accesable in application menu (Linux)
+    local_app_path = os.path.join(home_path, '.local/share/applications')
+    if os.path.exists(local_app_path) and os.path.isdir(local_app_path):
+        out_path = os.path.join(local_app_path, "lisa.desktop")
+        shutil.copy2(in_path, out_path)
+        
+        out_path_ha = os.path.join(local_app_path, "ha.desktop")
+        shutil.copy2(in_path_ha, out_path_ha)
+        
+        print "icon output path:"
+        print out_path, out_path_ha
 
-    for line in fileinput.input(out_path, inplace=True):
-        #coma on end makes no linebreak
-        print line.replace("@{LISA_PATH}", lisa_path),
-#        print(line.replace("@", lisa_path), end='')
-    #fi.close()
+        for line in fileinput.input(out_path, inplace=True):
+            #coma on end makes no linebreak
+            print line.replace("@{LISA_PATH}", lisa_path),
+            
+        for line in fileinput.input(out_path_ha, inplace=True):
+            #coma on end makes no linebreak
+            print line.replace("@{LISA_PATH}", lisa_path),
+    else:
+        print "Couldnt find $HOME/.local/share/applications/."
 
 
 def main():
